+++ dataprovider/foxsports/season.go
@@ -10,9 +10,9 @@ const (
 	PRESEASON
 )
 
-// SelectionId outputs the selection id used to fetch NFL matchups based on year, week, and season type
+// SegmentID outputs the selection id used to fetch NFL matchups based on year, week, and season type
 //
 // Example: https://api.foxsports.com/bifrost/v1/nfl/scoreboard/segment/2024-4-2?apikey=jE7yBJVRNAwdDesMgTzTXUUSx1It41Fq
-func (st SeasonType) SelectionId(year int32, week int32) string {
+func (st SeasonType) SegmentID(year int32, week int32) string {
 	return fmt.Sprintf("%d-%d-%d", year, week, st)
 }